moving content out of sequence folder to steps
