+++ src/main/steps/common/template.njk
@@ -1,0 +1,8 @@
+{% extends "./page.njk" %}
+
+{% block page_content %}
+  <h1 class="govuk-heading-l">{{ title }}</h1>
+  {% if line1 %}
+    <p class="govuk-body">{{ line1 }}</p>
+  {% endif %}
+{% endblock %}
