+++ src/main/steps/do-you-have-your-certificate/template.njk
