--- src/main/steps/sequence/template.njk
@@ -1,8 +1,0 @@
-{% extends "common/page.njk" %}
-
-{% block page_content %}
-  <h1 class="govuk-heading-l">{{ title }}</h1>
-  {% if line1 %}
-    <p class="govuk-body">{{ line1 }}</p>
-  {% endif %}
-{% endblock %}
