--- src/main/steps/sequence/your-details/template.njk
@@ -1,3 +1,0 @@
-{% extends "../template.njk" %}
-
-{% block backLink %}{% endblock %}
