+++ src/main/steps/your-details/template.njk
@@ -1,0 +1,3 @@
+{% extends "../common/template.njk" %}
+
+{% block backLink %}{% endblock %}
